pyproject.toml
[tool.poetry]
name = "langgraph-human-feedback"
version = "0.1.0"
description = ""
authors = ["Allan Ferreira <ferrerallan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
langgraph = "^0.2.74"
python-dotenv = "^1.0.1"
langchain = "^0.3.19"
langchain-core = "^0.3.39"
langgraph-checkpoint-sqlite = "^2.0.5"
openai = "^1.64.0"
langchain-openai = "^0.3.7"
langchain-community = "^0.3.18"
chromadb = "^0.6.3"
langchain-chroma = "^0.2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
